POC-90 Remove usage of Location class - using Geometry instead. Use double instead of BigDecimal because this is what the GeometryFactory uses. Expose lon/lat as double (not quoted in json). Test for nearby search for stop places.

+++ src/main/java/no/rutebanken/tiamat/repository/ifopt/StopPlaceRepositoryCustom.java
@@ -2,7 +2,6 @@ package no.rutebanken.tiamat.repository.ifopt;
 
 import uk.org.netex.netex.StopPlace;
 
-import java.math.BigDecimal;
 import java.util.List;
 
 
@@ -17,5 +16,5 @@ public interface StopPlaceRepositoryCustom {
      */
     StopPlace findStopPlaceDetailed(String id);
 
-    List<StopPlace> findStopPlacesWithin(BigDecimal xMin, BigDecimal yMin, BigDecimal xMax, BigDecimal yMax);
+    List<StopPlace> findStopPlacesWithin(double xMin, double yMin, double xMax, double yMax);
 }
